+++ CHANGELOG.md
@@ -6,15 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## v1.1.0
 ### Added
-
-### Fixed
-- Errors in readme.md
 - Weekly interval accepts constructor arguments
 - Weekly interval add magic every{day} method
 - Add BadMethodCallException for use by __call
 - Weekly interval add magic andEvery{day} method
 - Weekly interval add ofEveryWeek setters
 - Weekly interval add ofEvery{weeks}Week magic setter
+
+### Fixed
+- Errors in readme.md
 
 ## v1.0.0
 ### Added